+++ deployment/data-streams/changeset/reward-manager/general/call_claim_rewards_test.go
@@ -1,16 +1,15 @@
-package v0_5_0
+package general
 
 import (
 	"testing"
 
-	"github.com/ethereum/go-ethereum/common"
 	"github.com/stretchr/testify/require"
 
 	commonChangesets "github.com/smartcontractkit/chainlink/deployment/common/changeset"
 	"github.com/smartcontractkit/chainlink/deployment/data-streams/changeset/testutil"
 )
 
-func TestCallPayRecipients(t *testing.T) {
+func TestCallClaimRewards(t *testing.T) {
 	e := testutil.NewMemoryEnv(t, true)
 
 	e, rewardManagerAddr, _ := DeployRewardManagerAndLink(t, e)
@@ -20,18 +19,17 @@ func TestCallPayRecipients(t *testing.T) {
 
 	_, err := commonChangesets.Apply(t, e, nil,
 		commonChangesets.Configure(
-			PayRecipientsChangeset,
-			PayRecipientsConfig{
-				ConfigsByChain: map[uint64][]PayRecipients{
-					testutil.TestChain.Selector: {PayRecipients{
+			ClaimRewardsChangeset,
+			ClaimRewardsConfig{
+				ConfigsByChain: map[uint64][]ClaimRewards{
+					testutil.TestChain.Selector: {ClaimRewards{
 						RewardManagerAddress: rewardManagerAddr,
-						PoolID:               poolID,
-						Recipients:           []common.Address{},
+						PoolIDs:              [][32]byte{poolID},
 					}},
 				},
 			},
 		),
 	)
-	// Need Configured Fee Manager For PayRecipients Event
+	// Need Configured Fee Manager For ClaimRewards Event
 	require.NoError(t, err)
 }
