+++ deployment/data-streams/changeset/reward-manager/general/call_pay_recipients_test.go
@@ -1,15 +1,16 @@
-package v0_5_0
+package general
 
 import (
 	"testing"
 
+	"github.com/ethereum/go-ethereum/common"
 	"github.com/stretchr/testify/require"
 
 	commonChangesets "github.com/smartcontractkit/chainlink/deployment/common/changeset"
 	"github.com/smartcontractkit/chainlink/deployment/data-streams/changeset/testutil"
 )
 
-func TestCallClaimRewards(t *testing.T) {
+func TestCallPayRecipients(t *testing.T) {
 	e := testutil.NewMemoryEnv(t, true)
 
 	e, rewardManagerAddr, _ := DeployRewardManagerAndLink(t, e)
@@ -19,17 +20,18 @@ func TestCallClaimRewards(t *testing.T) {
 
 	_, err := commonChangesets.Apply(t, e, nil,
 		commonChangesets.Configure(
-			ClaimRewardsChangeset,
-			ClaimRewardsConfig{
-				ConfigsByChain: map[uint64][]ClaimRewards{
-					testutil.TestChain.Selector: {ClaimRewards{
+			PayRecipientsChangeset,
+			PayRecipientsConfig{
+				ConfigsByChain: map[uint64][]PayRecipients{
+					testutil.TestChain.Selector: {PayRecipients{
 						RewardManagerAddress: rewardManagerAddr,
-						PoolIDs:              [][32]byte{poolID},
+						PoolID:               poolID,
+						Recipients:           []common.Address{},
 					}},
 				},
 			},
 		),
 	)
-	// Need Configured Fee Manager For ClaimRewards Event
+	// Need Configured Fee Manager For PayRecipients Event
 	require.NoError(t, err)
 }
