+++ deployment/data-streams/changeset/reward-manager/general/call_set_fee_manager_test.go
@@ -1,4 +1,4 @@
-package v0_5_0
+package general
 
 import (
 	"testing"
