+++ deployment/data-streams/changeset/reward-manager/general/call_update_reward_recipients.go
@@ -1,4 +1,4 @@
-package v0_5_0
+package general
 
 import (
 	"errors"
