+++ deployment/data-streams/changeset/reward-manager/general/deploy_reward_manager_test.go
@@ -1,4 +1,4 @@
-package v0_5_0
+package general
 
 import (
 	"testing"
